Improved LUFactorization test
