+++ test/unit/CMakeLists.txt
@@ -13,6 +13,6 @@ add_unit_test(test_communicator_grid MPIRANKS 6 LIBS ${DLA_LIBS})
 add_unit_test(test_communicator_manager MPIRANKS 6 LIBS ${DLA_LIBS})
 add_unit_test(test_communicator_manager_init SRC test_communicator_manager.cpp MPIRANKS 6 DEFS -DCOMM_INITS_MPI LIBS ${DLA_LIBS})
 add_unit_test(test_distributed_matrix MPIRANKS 6 LIBS test_tools ${DLA_LIBS})
-add_unit_test(test_dla_interface MPIRANKS 6 LIBS test_tools ${DLA_LIBS})
 add_unit_test(test_dlai_cholesky_factorization MPIRANKS 6 LIBS test_tools ${DLA_LIBS})
 add_unit_test(test_dlai_matrix_multiplication MPIRANKS 6 LIBS test_tools ${DLA_LIBS})
+add_unit_test(test_dlai_lu_factorization MPIRANKS 6 LIBS test_tools ${DLA_LIBS})
